Add Goals Feature
Change userId from Integer/int to Long

+++ backend/src/main/java/com/kiwi/features/quests/controllers/QuestController.java
@@ -32,7 +32,7 @@ public class QuestController {
     public ResponseEntity<List<QuestDTO>> getActiveQuests(
             @AuthenticationPrincipal @NotNull UserDetails userDetails
     ) {
-        int userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
+        Long userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
                 .orElseThrow()
                 .getId();
 
@@ -44,7 +44,7 @@ public class QuestController {
     public ResponseEntity<List<QuestDTO>> getCompletedQuests(
             @AuthenticationPrincipal UserDetails userDetails
     ) {
-        int userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
+        Long userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
                 .orElseThrow()
                 .getId();
 
@@ -61,7 +61,7 @@ public class QuestController {
             @AuthenticationPrincipal UserDetails userDetails,
             @PathVariable int questId
     ) {
-        int userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
+        Long userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
                 .orElseThrow()
                 .getId();
 
@@ -78,7 +78,7 @@ public class QuestController {
             @AuthenticationPrincipal UserDetails userDetails,
             @PathVariable int subquestId
     ) {
-        int userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
+        Long userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
                 .orElseThrow()
                 .getId();
 
@@ -91,7 +91,7 @@ public class QuestController {
             @AuthenticationPrincipal UserDetails userDetails,
             @PathVariable int subquestId
     ) {
-        int userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
+        Long userId = usersService.getUserByEmail(new Email(userDetails.getUsername()))
                 .orElseThrow()
                 .getId();
